Move ZenformState to config package
